website/money-maker-1236/requirements.txt
Flask==0.12.2
gunicorn==19.7.1
google-cloud-datastore==1.2.0